Big_query_Chatbot/Part_B/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-bigquery==3.14.0
google-cloud-aiplatform==1.38.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
